backend/requirements.txt
fastapi
sqlalchemy
pydantic
pydantic_settings
bcrypt
passlib[bcrypt]
pyjwt